basic setup more file restructuring
