+++ src/pages/StorePortal/StorePortal.css
@@ -1,0 +1,6 @@
+.searchBarComponent {
+  height: 60vh;
+  width: 75vw;
+  margin-left: 12.5vw;
+  border: 5px solid red;
+}
